Переписал сервисный слой и контроллер для типов кредитов. Поднял маппинг и формирование отображения в контроллер, в сервисе оставил только бизнес-логику. Поправил дто для ответов с пагинацией: добавил общее число страниц.

+++ src/main/java/ru/trainee/creditmanager/controller/rest/v1/CreditTypeControllerV1.java
@@ -6,17 +6,25 @@ import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.persistence.EntityExistsException;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.AllArgsConstructor;
+import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import ru.trainee.creditmanager.dto.PageableListResponseDTO;
 import ru.trainee.creditmanager.dto.creditType.CreditTypeCreateDTO;
 import ru.trainee.creditmanager.dto.creditType.CreditTypeResponseDetailDTO;
 import ru.trainee.creditmanager.dto.creditType.CreditTypeResponseShortDTO;
 import ru.trainee.creditmanager.dto.creditType.CreditTypeUpdateDTO;
+import ru.trainee.creditmanager.entity.CreditType;
+import ru.trainee.creditmanager.mapper.creditType.CreditTypeMapper;
 import ru.trainee.creditmanager.service.CreditTypeService;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @RestController
@@ -26,6 +34,7 @@ import java.util.UUID;
 public class CreditTypeControllerV1 {
 
     private final CreditTypeService creditTypeService;
+    private final CreditTypeMapper creditTypeMapper;
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
@@ -41,7 +50,16 @@ public class CreditTypeControllerV1 {
             {@Content(mediaType = "application/json", schema =
             @Schema(implementation = CreditTypeResponseDetailDTO.class))})
     public CreditTypeResponseDetailDTO create(@RequestBody CreditTypeCreateDTO dto) {
-        return creditTypeService.create(dto);
+
+        Optional<CreditType> creditTypeOptional = creditTypeService.findByName(dto.getName());
+        if(creditTypeOptional.isEmpty()){
+            return creditTypeMapper.toCreditTypeDetailDto(
+                    creditTypeService.create(
+                            creditTypeMapper.toCreditTypeEntity(dto))
+            );
+        } else {
+            throw new EntityExistsException("Credit type already exixst.");
+        }
     }
 
 
@@ -52,11 +70,18 @@ public class CreditTypeControllerV1 {
                     Список содержит краткие карточки типов кредитов.
                     """
     )
-    public List<CreditTypeResponseShortDTO> readAll(
-            @RequestParam(required = false, defaultValue = "0") int page,
-            @RequestParam(required = false, defaultValue = "10") int size
-    ) {
-        return creditTypeService.readAll(PageRequest.of(page, size));
+    public PageableListResponseDTO<CreditTypeResponseShortDTO> readAll(Pageable p){
+
+        Page<CreditType> page = creditTypeService.readAll(p);
+
+        return new PageableListResponseDTO<>(
+                page.getNumber(),
+                page.getSize(),
+                page.getTotalElements(),
+                page.getTotalPages(),
+                page.getContent().stream()
+                        .map(creditTypeMapper::toCreditTypeShortDto)
+                        .toList());
     }
 
 
@@ -70,8 +95,12 @@ public class CreditTypeControllerV1 {
                               """
     )
     public CreditTypeResponseDetailDTO findById(@PathVariable UUID id) {
-
-        return creditTypeService.findById(id);
+        Optional<CreditType> creditTypeOptional = creditTypeService.findById(id);
+        if(creditTypeOptional.isPresent()){
+            return creditTypeMapper.toCreditTypeDetailDto(creditTypeOptional.get());
+        } else {
+            throw new EntityNotFoundException("Credit type " + id + " not found! Check id.");
+        }
     }
 
 
@@ -86,8 +115,11 @@ public class CreditTypeControllerV1 {
                               """
     )
     @ResponseStatus(HttpStatus.OK)
-    public CreditTypeResponseDetailDTO update(@RequestBody CreditTypeUpdateDTO creditType){
-        return creditTypeService.update(creditType);
+    public CreditTypeResponseDetailDTO update(@RequestBody CreditTypeCreateDTO dto){
+        return creditTypeMapper.toCreditTypeDetailDto(
+                creditTypeService.update(
+                        creditTypeMapper.toCreditTypeEntity(dto))
+        );
     }
 
 
@@ -100,6 +132,9 @@ public class CreditTypeControllerV1 {
                               """
     )
     public void delete(@PathVariable UUID id) {
+        if(creditTypeService.findById(id).isEmpty()){
+            throw new EntityNotFoundException("Credit type " + id + " not found! Check id.");
+        }
         creditTypeService.delete(id);
     }
 }
